+++ src/aui-scheduler/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-2074](https://issues.liferay.com/browse/AUI-2074) Replace direct call to "filter" function with "Array.filter"
 * [AUI-2061](https://issues.liferay.com/browse/AUI-2061) Low performance on countDays method when using huge dates.
 * [AUI-2070](https://issues.liferay.com/browse/AUI-2070) Replace direct call to "filter" function with "Array.filter"