+++ src/aui-scheduler/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-* [AUI-2074](https://issues.liferay.com/browse/AUI-2074) Replace direct call to "filter" function with "Array.filter"
+* [AUI-2074](https://issues.liferay.com/browse/AUI-2074) Cannot view all day events in Day view
 * [AUI-2061](https://issues.liferay.com/browse/AUI-2061) Low performance on countDays method when using huge dates.
 * [AUI-2070](https://issues.liferay.com/browse/AUI-2070) Replace direct call to "filter" function with "Array.filter"