+++ src/aui-scheduler/HISTORY.md
@@ -4,6 +4,10 @@
 
 ## @VERSION@
 
+No registries yet.
+
+## [3.1.0](https://github.com/liferay/alloy-ui/releases/tag/3.1.0)
+
 * [AUI-2281](https://issues.liferay.com/browse/AUI-2106) Use arbitrary time to display the red line in day view.
 * [AUI-2106](https://issues.liferay.com/browse/AUI-2106) It should be possible to not show the scheduler's toolbar.
 * [AUI-2080](https://issues.liferay.com/browse/AUI-2080) Agenda view displays too many events if start and end dates are too far away.